terraform/modules/cloudlab/cloud_functions/welcome-email/requirements.txt
# Copyright 2023 Google. This software is provided as-is, without warranty
# or representation for any use or purpose. Your use of it is subject to
# your agreement with Google.
google-cloud-resource-manager
google.api_core
google.cloud
